# SupplyInventory/Application/Controllers/PalletRack/PalletRackController.go
package controllers

import (
	"net/http"
	"strconv"

	requestsHelper "go_inventory/Helpers/RequestsHelper"
	middlewares "go_inventory/SupplyInventory/Application/Middlewares"
	palletRackRequests "go_inventory/SupplyInventory/Application/Requests/PalletRack"
	palletrack "go_inventory/SupplyInventory/Application/Services/PalletRack"

	"github.com/gin-gonic/gin"
)

type PalletRackController struct {
	palletRackService palletrack.PalletRackService
}

func NewPalletRackController(palletRackService palletrack.PalletRackService) *PalletRackController {
	return &PalletRackController{palletRackService: palletRackService}
}

func (controller *PalletRackController) RegisterPalletRack(group *gin.RouterGroup) {
	group.POST("/", controller.createPalletRack)
	group.GET("/", controller.listRacks)
	group.GET("/:id", controller.FindRackById)
	group.DELETE("/:id", controller.DeleteRack)
}

func (controller *PalletRackController) parsePage(c *gin.Context) int {
	pageStr := c.Query("page")
	if pageStr == "" {
		return 1
	}

	page, err := strconv.Atoi(pageStr)
	if err != nil || page <= 0 {
		return 1
	}

	return page
}

func (controller *PalletRackController) parseLimit(c *gin.Context) int {
	limitStr := c.Query("limit")
	if limitStr == "" {
		return 10
	}

	limit, err := strconv.Atoi(limitStr)
	if err != nil || limit <= 0 || limit > 100 {
		return 10
	}

	return limit
}

// @Summary Create Pallet Racks
// @Tags Pallet Racks
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param PalletRack body paletrack.PalletRackRequest true "Palletized Product"
// @Success 200 {object} entities.PalletRackEntity
// @Failure 422 {object} map[string]string
// @Router /api/v1/racks [post]
func (controller *PalletRackController) createPalletRack(c *gin.Context) {
	var req palletRackRequests.PalletRackRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusUnprocessableEntity, gin.H{"error": requestsHelper.FormatValidationErrors(err)})
		return
	}

	inventoryID := middlewares.GetInventoryID(c)
	newPalletRack, err := controller.palletRackService.Create(req.Name, req.Location, req.TotalCapacity, inventoryID)

	if newPalletRack == nil {
		c.JSON(http.StatusUnprocessableEntity, gin.H{"message": err.Error()})
		return
	}

	c.JSON(http.StatusCreated, newPalletRack)
}

// @Summary List Pallet Racks
// @Tags Pallet Racks
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param page query int false "Page number" default(1)
// @Param limit query int false "Items per page" default(10)
// @Success 200 {object} apiContracts.PaginatedRacksResponse
// @Failure 404 {object} map[string]string
// @Router /api/v1/racks [get]
func (controller *PalletRackController) listRacks(c *gin.Context) {
	inventoryID := middlewares.GetInventoryID(c)
	page := controller.parsePage(c)
	limit := controller.parseLimit(c)

	racks, err := controller.palletRackService.ListRacks(&inventoryID, page, limit)
	if err != nil {
		c.JSON(http.StatusNotFound, gin.H{})
		return
	}

	c.JSON(http.StatusOK, racks)
}

// @Summary Get Rack
// @Tags Pallet Racks
// @Accept json
// @Produce json
// @Security BearerAuth
// @Success 200 {object} entities.PalletRackEntity
// @Failure 404 {object} map[string]string
// @Failure 422 {object} map[string]string
// @Param id path int true "ID do Rack"
// @Router /api/v1/racks/{id} [get]
func (pc *PalletRackController) FindRackById(c *gin.Context) {
	id, err := requestsHelper.GetIDParam(c, "id")
	if err != nil {
		c.JSON(http.StatusUnprocessableEntity, gin.H{"message": "ID inválido"})
		return
	}

	pallet, appErr := pc.palletRackService.FindPalletById(id)
	if pallet == nil {
		c.JSON(appErr.ErrorCode(), gin.H{"message": appErr.Error()})
		return
	}

	c.JSON(http.StatusOK, pallet)
}

// @Summary Delete Rack
// @Tags Pallet Racks
// @Accept json
// @Produce json
// @Security BearerAuth
// @Success 204 "Delete successful"
// @Failure 404 {object} map[string]string
// @Failure 422 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Param id path int true "ID do Rack"
// @Router /api/v1/racks/{id} [delete]
func (pc *PalletRackController) DeleteRack(c *gin.Context) {
	id, err := requestsHelper.GetIDParam(c, "id")
	if err != nil {
		c.JSON(http.StatusUnprocessableEntity, gin.H{"message": "ID inválido"})
		return
	}

	_, appErr := pc.palletRackService.DeleteRack(id)
	if appErr != nil {
		c.JSON(appErr.ErrorCode(), gin.H{"message": appErr.Error()})
		return
	}

	c.Status(http.StatusOK)
}