# docs/swagger.yaml
definitions:
...
  apiContracts.PaginatedRacksResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/apiContracts.TransformedRack'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  apiContracts.TransformedRack:
    properties:
      id:
        type: integer
      location:
        type: string
      name:
        type: string
      pallets:
        items:
          $ref: '#/definitions/entities.PalletEntity'
        type: array
      percetageUsed:
        format: float64
        type: number
      totalCapacity:
        type: integer
    type: object
...
  entities.PalletEntity:
    properties:
      created_at:
        type: string
      id:
        type: integer
      inventory_id:
        type: integer
      name:
        type: string
      palletRackId:
        type: integer
      palletRackName:
        type: string
      palletizedProduct:
        items:
          $ref: '#/definitions/entities.PalletizedProductEntity'
        type: array
      qr_code:
        type: string
      qr_code_url:
        type: string
      updated_at:
        type: string
    required:
    - inventory_id
    - name
    - palletRackId
    - palletRackName
    type: object
  entities.PalletRackEntity:
    properties:
      created_at:
        type: string
      id:
        type: integer
      inventory_id:
        type: integer
      location:
        type: string
      name:
        type: string
      pallets:
        items:
          $ref: '#/definitions/entities.PalletEntity'
        type: array
      total_capacity:
        type: integer
      updated_at:
        type: string
    required:
    - inventory_id
    - location
    - name
    - total_capacity
    type: object
  entities.PalletizedProductEntity:
    properties:
      created_at:
        type: string
      ean:
        type: integer
      id:
        type: integer
      inventory_id:
        type: integer
      palletID:
        type: integer
      quantity:
        type: integer
      updated_at:
        type: string
    required:
    - ean
    - quantity
    type: object
...
  paletrack.PalletRackRequest:
    properties:
      location:
        type: string
      name:
        type: string
      totalCapacity:
        minimum: 1
        type: integer
    required:
    - location
    - name
    - totalCapacity
    type: object
  pallet.PalletRequest:
    properties:
      name:
        type: string
      palletRackId:
        type: integer
    required:
    - name
    - palletRackId
    type: object
  palletizedproduct.PalletizedProductRequest:
    properties:
      ean:
        type: integer
      quantity:
        type: integer
    required:
    - ean
    - quantity
    type: object
...
host: localhost:8000
info:
  contact:
    email: joaoferrazp@gmail.com
    name: API Support
  description: It is system to manage pallets and related products.
  title: Go Inventory API
  version: "1.0"
paths:
  /api/v1/admin/racks:
    get:
      consumes:
      - application/json
      parameters:
      - description: Inventory ID to filter
        in: query
        name: inventory_id
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiContracts.PaginatedRacksResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Admin List Pallet Racks
      tags:
      - Admin Pallet Racks
...
  /api/v1/pallet/products/{palletId}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID do pallet
        in: path
        name: palletId
        required: true
        type: integer
      - description: Palletized Product
        in: body
        name: pallet
        required: true
        schema:
          $ref: '#/definitions/palletizedproduct.PalletizedProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.PalletizedProductEntity'
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a product to a pallet
      tags:
      - Palletized products
  /api/v1/pallet/products/{palletId}/{productsEan}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID do pallet
        in: path
        name: palletId
        required: true
        type: integer
      - description: EAN do produto
        in: path
        name: productsEan
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete product from pallet
      tags:
      - Palletized products
  /api/v1/pallets:
    get:
      consumes:
      - application/json
      parameters:
      - description: Filter by pallet rack ID
        in: query
        name: palletRackId
        type: integer
      - description: Filter by product EAN
        in: query
        name: ean
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.PalletEntity'
            type: array
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: List pallets
      tags:
      - Pallets
    post:
      consumes:
      - application/json
      parameters:
      - description: Palletized Product
        in: body
        name: pallet
        required: true
        schema:
          $ref: '#/definitions/pallet.PalletRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.PalletEntity'
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create Pallet
      tags:
      - Pallets
  /api/v1/pallets/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Palletized Product
        in: body
        name: pallet
        required: true
        schema:
          $ref: '#/definitions/pallet.PalletRequest'
      - description: ID do pallet
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.PalletEntity'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Pallet
      tags:
      - Pallets
    get:
      consumes:
      - application/json
      parameters:
      - description: ID do pallet
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.PalletEntity'
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Get pallet by ID
      tags:
      - Pallets
    patch:
      consumes:
      - application/json
      parameters:
      - description: Palletized Product
        in: body
        name: pallet
        required: true
        schema:
          $ref: '#/definitions/pallet.PalletRequest'
      - description: ID do pallet
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.PalletEntity'
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update Pallet
      tags:
      - Pallets
...
  /api/v1/racks:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiContracts.PaginatedRacksResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List Pallet Racks
      tags:
      - Pallet Racks
    post:
      consumes:
      - application/json
      parameters:
      - description: Palletized Product
        in: body
        name: PalletRack
        required: true
        schema:
          $ref: '#/definitions/paletrack.PalletRackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.PalletRackEntity'
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create Pallet Racks
      tags:
      - Pallet Racks
  /api/v1/racks/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID do Rack
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Delete successful
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Rack
      tags:
      - Pallet Racks
    get:
      consumes:
      - application/json
      parameters:
      - description: ID do Rack
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.PalletRackEntity'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Rack
      tags:
      - Pallet Racks
...
securityDefinitions:
  BearerAuth:
    description: 'Digite: "Bearer {token}" para se autenticar.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"